+++ lib/core/crypto/eosdart/eosdart.dart
@@ -14,6 +14,5 @@ export 'src/models/block_header_state.dart';
 export 'src/models/node_info.dart';
 export 'src/models/primary_wrapper.dart';
 export 'src/models/transaction.dart';
-export 'src/models/transaction.dart';
 export 'src/numeric.dart';
 export 'src/serialize.dart';
