+++ lib/core/crypto/seeds_esr/seeds_esr.dart
@@ -1,4 +1,5 @@
 import 'package:async/async.dart';
+import 'package:get_it/get_it.dart';
 import 'package:hypha_wallet/core/crypto/dart_esr/src/encoding_options.dart';
 import 'package:hypha_wallet/core/crypto/dart_esr/src/models/authorization.dart';
 import 'package:hypha_wallet/core/crypto/dart_esr/src/models/esr_action.dart';
@@ -7,6 +8,7 @@ import 'package:hypha_wallet/core/crypto/seeds_esr/eos_transaction.dart';
 import 'package:hypha_wallet/core/crypto/seeds_esr/scan_qr_code_result_data.dart';
 import 'package:hypha_wallet/core/logging/log_helper.dart';
 import 'package:hypha_wallet/core/network/api/endpoints.dart';
+import 'package:hypha_wallet/core/network/api/remote_config_serivice.dart';
 // import 'package:hypha_wallet/core/crypto/eosdart/src/models/action.dart';
 
 class SeedsESR {
@@ -24,12 +26,6 @@ class SeedsESR {
     actions = await manager.fetchActions(account: account);
   }
 
-  // TODO(n13): Remove this method and replace it with something more sensible. This is using the already resolved
-  // ESR object and creating a lightweight action data object out of the first action, so be passed around to
-  // other components.
-  // Better ways to do that
-  // Pass around the whole ESR object, or an Action object.
-  // instead of canProcess, have an isValid accessor on the ESR and handle this case in the mappers.
   Result<ScanQrCodeResultData> processResolvedRequest() {
     final EOSTransaction eosTransaction = EOSTransaction.fromESRActionsList(actions);
     if (eosTransaction.isValid) {
@@ -45,7 +41,8 @@ class SeedsESR {
 extension TelosSigningManager on SigningRequestManager {
   static SigningRequestManager from(String? uri) {
     return SigningRequestManager.from(uri,
-        options: defaultSigningRequestEncodingOptions(nodeUrl: Endpoints.pushTransactionNodeUrl()));
+        options:
+            defaultSigningRequestEncodingOptions(nodeUrl: GetIt.I<RemoteConfigService>().pushTransactionNodeUrl()));
   }
 
   Future<List<ESRAction>> fetchActions({String? account, String permission = 'active'}) async {
