+++ lib/core/network/api/eos_service.dart
@@ -2,11 +2,12 @@ import 'dart:async';
 
 import 'package:async/async.dart';
 import 'package:dio/dio.dart' as dio;
+import 'package:get_it/get_it.dart';
 import 'package:hypha_wallet/core/crypto/eosdart/eosdart.dart';
 import 'package:hypha_wallet/core/crypto/seeds_esr/eos_transaction.dart';
 import 'package:hypha_wallet/core/local/models/user_auth_data.dart';
 import 'package:hypha_wallet/core/local/services/secure_storage_service.dart';
-import 'package:hypha_wallet/core/network/api/endpoints.dart';
+import 'package:hypha_wallet/core/network/api/remote_config_serivice.dart';
 
 class EOSService {
   final EOSClient eosClient;
@@ -33,7 +34,7 @@ class EOSService {
     final UserAuthData? userAuthData = await secureStorageService.getUserAuthData();
 
     final eosClient = EOSClient(
-      baseUrl: Endpoints.pushTransactionNodeUrl(),
+      baseUrl: GetIt.I<RemoteConfigService>().pushTransactionNodeUrl(),
       privateKeys: [userAuthData!.eOSPrivateKey.toString()],
       version: 'v1',
     );
